Add document collection ids to links payload for Email Alert API
This is needed to support
https://github.com/alphagov/email-alert-api/pull/1803

When a user signs up to alerts on a document collection, research has shown an
expectation of being notified when changes are made to documents within that collection.

The message that publishing API puts onto rabbit contains the expanded links (see eg here
https://deploy.integration.publishing.service.gov.uk/job/run-rake-task/305143/console) but
email alert service only forwards some of this to Email Alert API:

  - To support topic taxonomy based emails, this application creates a custom `taxon_tree`
attribute populated from content_ids in the taxon attribute of expanded links.
  - Following that pattern, this commit adds a document collection attribute, populated
with content ids plucked from the document collection attribute of expanded links