Add reverse links to the links payload for email alert API
When a major change is published, publishing API's downstream workers put a message on rabbit
that contains the expanded links[1] (see [2] for example output). But this application does not
currently send the expanded links onwards to Email Alert API.
We have a user need for document collection subscribers to be notified when a piece of content within that collection
is updated. This relationship is stored as a reverse link[3] in the expanded links for the document.

Available translations are added in to the expanded links[4], but there is no sense in email alert API
knowing about that hence their removal on line 47

[1] https://github.com/alphagov/publishing-api/blob/main/app/workers/downstream_live_worker.rb#L39
[2] https://deploy.integration.publishing.service.gov.uk/job/run-rake-task/305143/console
[3] https://github.com/alphagov/publishing-api/blob/main/lib/expansion_rules.rb#L43-L56
[4] https://github.com/alphagov/publishing-api/blob/4ba7062e7b8d525d2d0011d9b91aad801066aedf/app/presenters/queries/expanded_link_set.rb#L69