Add a single entrypoint that runs all consumers.
On Kubernetes it's simpler and more efficient to run the four consumer
processes in the same Pod than to run a separate Deployment for each.

The simplest way to do this is to fork and have the parent process exit
if any of the child processes exit.

The main motivation for this change is to avoid having to add weirdness
to the Prometheus scrape config and run [prometheus_exporter] on
multiple ports in the same Pod, which is what we'd have to do if we were
to keep the application as-is.

Avoiding the (mis)use of Rake to start the application simplifies the
startup process and makes it easier to understand the control flow
during initialisation. For example it turns out that Bootsnap was being
initialised much later than it's supposed to be, even though from
reading the source code it looked like the first thing called on
startup.

Alternatives considered

If we were to keep the existing model and run four Rake containers in
the same Pod then we'd need to assign four separate ports for the
Prometheus metrics in order for them not to clash, which just seems
overly complex vs. just forking a process for each consumer loop. If
[prometheus_exporter] could be made to run as a sidecar container
alongside the four Rake containers then we could have a single port for
`/metrics` and this would be a viable (though somewhat more complex)
option.

The other alternative of running a separate Deployment for each message
type would use 3-4x the resources, given the very low rates of messages
on three out of the four queues. So that'd be worse for running costs
and CO₂e, as well as being slightly more operationally complex.

[prometheus_exporter]: https://github.com/discourse/prometheus_exporter

Co-authored-by: David Mays <david.mays@digital.cabinet-office.gov.uk>